[synthetics] Extract renderResults function from command
